Started Adding Map System
Added MapNodes for finding path between mesh network of stars and the stars representde by those nodes, as well as the angle that the ship must jump to.

Added MapLink to display what nodes are connected (drawn at the beginning of runtime). Can be used to display the current planned path (by changing color).

To do:
I need to change MapLink from being a LineRenderer to a UILineRenderer (bitbucket). Right now it has to update every frame and it just looks a little shoddy.